ai interview1/interview-system/requirements.txt
flask==2.0.1
opencv-python==4.5.3.56
pytesseract==0.3.8
python-dotenv==0.19.0
SpeechRecognition==3.8.1
pyttsx3==2.90
mediapipe==0.8.9.1
PyPDF2==1.26.0